Rearange code
